+++ Pod/Classes/Core/ARCocoaConstantsModule.m
@@ -1,3 +1,5 @@
+@import ARKit;
+
 #import "ARCocoaConstantsModule.h"
 
 #import <UIKit/UIKit.h>
@@ -13,9 +15,17 @@ RCT_EXPORT_MODULE();
 
 - (NSDictionary *)constantsToExport;
 {
+    BOOL isAREnabled = NO;
+
+    if (@available(iOS 11.3, *)) {
+        isAREnabled = [ARWorldTrackingConfiguration isSupported];
+    }
+
     return @{
         @"UIApplicationOpenSettingsURLString": UIApplicationOpenSettingsURLString,
         @"LocalTimeZone": [[NSTimeZone localTimeZone] name],
+        @"CurrentLocale": [[NSLocale currentLocale] localeIdentifier],
+        @"AREnabled": @(isAREnabled),
     };
 }