+++ src/lib/Components/Consignments/Screens/__tests__/Overview-local-storage-tests.tsx
@@ -5,6 +5,7 @@ AsyncStorage.removeItem = jest.fn()
 
 import Overview from "../Overview"
 
+jest.mock("@react-native-community/cameraroll", () => jest.fn())
 jest.mock("../../Submission/create", () => jest.fn())
 jest.mock("../../Submission/update", () => jest.fn())