+++ src/lib/Scenes/Artwork/Artwork.tsx
@@ -1,10 +1,16 @@
+import { Flex, Theme } from "@artsy/palette"
 import { Artwork_artwork } from "__generated__/Artwork_artwork.graphql"
 import { ArtworkQuery } from "__generated__/ArtworkQuery.graphql"
+import Separator from "lib/Components/Separator"
 import { defaultEnvironment } from "lib/relay/createEnvironment"
 import renderWithLoadProgress from "lib/utils/renderWithLoadProgress"
 import React from "react"
-import { Text, View } from "react-native"
+import { View } from "react-native"
 import { createFragmentContainer, graphql, QueryRenderer } from "react-relay"
+import { ArtworkActionsFragmentContainer as ArtworkActions } from "./Components/ArtworkActions"
+import { ArtworkAvailabilityFragmentContainer as ArtworkAvailability } from "./Components/ArtworkAvailability"
+import { ArtworkTombstoneFragmentContainer as ArtworkTombstone } from "./Components/ArtworkTombstone"
+import { SellerInfoFragmentContainer as SellerInfo } from "./Components/SellerInfo"
 
 interface Props {
   artwork: Artwork_artwork
@@ -12,10 +18,22 @@ interface Props {
 
 export class Artwork extends React.Component<Props> {
   render() {
+    const { artwork } = this.props
     return (
-      <View style={{ backgroundColor: "red" }}>
-        <Text>{this.props.artwork.title}</Text>
-      </View>
+      <Theme>
+        <View>
+          <Flex width="100%" style={{ backgroundColor: "gray" }} height={340} />
+          <Flex alignItems="center" mt={2}>
+            <ArtworkActions artwork={artwork} />
+            <ArtworkTombstone artwork={artwork} />
+          </Flex>
+          <Separator />
+          <Flex width="100%">
+            <ArtworkAvailability artwork={artwork} />
+            <SellerInfo artwork={artwork} />
+          </Flex>
+        </View>
+      </Theme>
     )
   }
 }
@@ -23,7 +41,10 @@ export class Artwork extends React.Component<Props> {
 export const ArtworkContainer = createFragmentContainer(Artwork, {
   artwork: graphql`
     fragment Artwork_artwork on Artwork {
-      title
+      ...ArtworkTombstone_artwork
+      ...ArtworkActions_artwork
+      ...ArtworkAvailability_artwork
+      ...SellerInfo_artwork
     }
   `,
 })