+++ src/lib/Scenes/Artwork/Components/__stories__/ArtworkAvailability.story.tsx
@@ -1,0 +1,7 @@
+import { storiesOf } from "@storybook/react-native"
+import React from "react"
+import { ArtworkAvailability } from "../ArtworkAvailability"
+
+storiesOf("Artwork/Components").add("Availability", () => {
+  return <ArtworkAvailability artwork={{ __id: "Blah", _id: "HI", availability: "On Loan", " $refType": null }} />
+})