+++ src/lib/Scenes/Artwork/__tests__/Artwork-tests.tsx
@@ -1,0 +1,11 @@
+import { shallow } from "enzyme"
+import React from "react"
+import { Artwork } from "../Artwork"
+import { ArtworkActions } from "../Components/ArtworkActions"
+
+describe("Artwork", () => {
+  it("renders a snapshot", () => {
+    const component = shallow(<Artwork artwork={{ " $fragmentRefs": null, " $refType": null }} />)
+    expect(component.find(ArtworkActions).length).toEqual(1)
+  })
+})