+++ CHANGELOG.md
@@ -16,6 +16,28 @@
 
 ### Master
 
+### 1.11.4
+
+- Adds a non-interactive ArtworkActions component - sweir27
+- Adds basic ArtworkAvailability component - ashkan18
+- Exposes AREnabled and a user's current locale - sweir27
+- Adds seller info section with SellerInfo component - ashkan18
+- Adds ArtworkTombstone component - lilyfromseattle
+
+### 1.11.3
+
+- Updated background color to demonstrate beta flow - luc
+
+### 1.11.2
+
+- Moves from RN CameraRoll to @react-native-community/cameraroll - ash
+
+### 1.11.1
+
+- Fixes consignment photo selection - ash
+
+### 1.11.0
+
 - Bootstrap Artwork component view controller - alloy & david
 
 ### 1.10.1